AppLinks and Directories Resource enhancements

+++ src/main/java/de/aservo/confapi/jira/service/DirectoriesServiceImpl.java
@@ -10,19 +10,18 @@ import de.aservo.confapi.commons.model.AbstractDirectoryBean;
 import de.aservo.confapi.commons.model.DirectoriesBean;
 import de.aservo.confapi.commons.model.DirectoryCrowdBean;
 import de.aservo.confapi.commons.service.api.DirectoriesService;
-import de.aservo.confapi.commons.util.BeanValidationUtil;
 import de.aservo.confapi.jira.model.util.DirectoryBeanUtil;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.stereotype.Component;
 
 import javax.inject.Inject;
-import javax.validation.constraints.NotNull;
-import java.util.Collections;
+import java.util.ArrayList;
+import java.util.List;
 import java.util.Optional;
-import java.util.stream.Collectors;
 
 import static com.google.common.base.Preconditions.checkNotNull;
+import static java.lang.String.format;
 
 @Component
 @ExportAsService(DirectoriesService.class)
@@ -38,43 +37,68 @@ public class DirectoriesServiceImpl implements DirectoriesService {
         this.crowdDirectoryService = checkNotNull(crowdDirectoryService);
     }
 
+    @Override
     public DirectoriesBean getDirectories() {
-        return new DirectoriesBean(crowdDirectoryService.findAllDirectories().stream().map(DirectoryBeanUtil::toDirectoryBean).collect(Collectors.toList()));
+        List<AbstractDirectoryBean> beans = new ArrayList<>();
+        for (Directory directory : crowdDirectoryService.findAllDirectories()) {
+            AbstractDirectoryBean crowdBean;
+            crowdBean = DirectoryBeanUtil.toDirectoryBean(directory);
+            beans.add(crowdBean);
+        }
+        return new DirectoriesBean(beans);
     }
 
-    public DirectoriesBean setDirectories(
-            DirectoriesBean directoriesBean, boolean testConnection) {
+    @Override
+    public DirectoriesBean setDirectories(DirectoriesBean directoriesBean, boolean testConnection) {
+        directoriesBean.getDirectories().forEach(directoryBaseBean -> {
+            if (directoryBaseBean instanceof DirectoryCrowdBean) {
 
-        DirectoryCrowdBean directoryBean = (DirectoryCrowdBean) directoriesBean.getDirectories().iterator().next();
+                //preps and validation
+                DirectoryCrowdBean crowdBean = (DirectoryCrowdBean)directoryBaseBean;
+                Directory directory = createDirectoryConfig(crowdBean, testConnection);
 
-        //preps and validation
-        BeanValidationUtil.validate(directoryBean);
-        Directory directory = DirectoryBeanUtil.toDirectory(directoryBean);
-        String directoryName = directoryBean.getName();
+                //check if directory exists already and if yes, remove it
+                Optional<Directory> presentDirectory = crowdDirectoryService.findAllDirectories().stream()
+                        .filter(dir -> dir.getName().equals(directory.getName())).findFirst();
+                if (presentDirectory.isPresent()) {
+                    Directory presentDir = presentDirectory.get();
+                    log.info("removing existing user directory configuration '{}' before adding new configuration '{}'", presentDir.getName(), directory.getName());
+                    try {
+                        crowdDirectoryService.removeDirectory(presentDir.getId());
+                    } catch (DirectoryCurrentlySynchronisingException e) {
+                        throw new InternalServerErrorException(e.getMessage());
+                    }
+                }
+
+                //add new directory
+                crowdDirectoryService.addDirectory(directory);
+            } else {
+                throw new InternalServerErrorException(format("Setting directory type '%s' is not supported (yet)", directoryBaseBean.getClass()));
+            }
+        });
+        return getDirectories();
+    }
+
+    @Override
+    public AbstractDirectoryBean addDirectory(AbstractDirectoryBean abstractDirectoryBean, boolean testConnection) {
+        if (abstractDirectoryBean instanceof DirectoryCrowdBean) {
+            DirectoryCrowdBean crowdBean = (DirectoryCrowdBean)abstractDirectoryBean;
+            Directory directory = createDirectoryConfig(crowdBean, testConnection);
+            Directory addedDirectory = crowdDirectoryService.addDirectory(directory);
+            return DirectoryBeanUtil.toDirectoryBean(addedDirectory);
+        } else {
+            throw new InternalServerErrorException(format("Adding directory type '%s' is not supported (yet)", abstractDirectoryBean.getClass()));
+        }
+    }
+
+    private Directory createDirectoryConfig(DirectoryCrowdBean crowdBean, boolean testConnection) {
+        Directory directory = DirectoryBeanUtil.toDirectory(crowdBean);
+        String directoryName = crowdBean.getName();
         if (testConnection) {
             log.debug("testing user directory connection for {}", directoryName);
             crowdDirectoryService.testConnection(directory);
         }
-
-        //check if directory exists already and if yes, remove it
-        Optional<Directory> presentDirectory = crowdDirectoryService.findAllDirectories().stream().filter(dir -> dir.getName().equals(directory.getName())).findFirst();
-        if (presentDirectory.isPresent()) {
-            Directory presentDir = presentDirectory.get();
-            log.info("removing existing user directory configuration '{}' before adding new configuration '{}'", presentDir.getName(), directory.getName());
-            try {
-                crowdDirectoryService.removeDirectory(presentDir.getId());
-            } catch (DirectoryCurrentlySynchronisingException e) {
-                throw new InternalServerErrorException(e.getMessage());
-            }
-        }
-
-        //add new directory
-        return new DirectoriesBean(Collections.singletonList(DirectoryBeanUtil.toDirectoryBean(crowdDirectoryService.addDirectory(directory))));
-    }
-
-    @Override
-    public AbstractDirectoryBean addDirectory(@NotNull AbstractDirectoryBean abstractDirectoryBean, boolean b) {
-        throw new UnsupportedOperationException("Not implemented");
+        return directory;
     }
 
 }
